+++ docs/changes.rst
@@ -6,6 +6,10 @@ Pending
 
 * Removed outdated third-party panels from the list.
 * Avoided the unnecessary work of recursively quoting SQL parameters.
+* Postponed context process in templates panel to include lazy evaluated
+  content.
+* Fixed template panel to avoid evaluating ``LazyObject`` when not already
+  evaluated.
 
 4.2.0 (2023-08-10)
 ------------------